cv_threat_detection/requirements.txt
ultralytics>=8.3.0
opencv-python
numpy
pillow
pyyaml
matplotlib
psutil
streamlit
